refactor(DB): 修改AllocateFileNumber返回类型为uint64_t
- 将DBImpl::AllocateFileNumber()返回类型从int改为uint64_t
- 更新实现中使用manifest_state_.next_file_number替代next_file_number_
- 头文件中同步修改函数声明返回类型为uint64_t

+++ include/DBImpl.h
@@ -50,7 +50,7 @@ class DBImpl {
         void LoadSSTables();
 
         // 管理next_file_number_
-        int AllocateFileNumber();
+        uint64_t AllocateFileNumber();
         // 负责计算并设置next_file_number_
         void InitNextFileNumberFromDisk();
